+++ include/openrave/trajectory.h
@@ -202,6 +202,7 @@ private:
 
     virtual void Clone(InterfaceBaseConstPtr preference, int cloningoptions);
 
+    virtual void SetAffineVelocity(const Vector& maxAffineTranslationVel, dReal maxAffineRotationQuatVel);
 private:
 
     /// \brief Linear interpolation using the maximum joint velocities for timing.